VRAM_Hへオーバーレイ階層を移動するので、static変数をu8->u32へ。

+++ prog/src/battle/btl_adapter.c
@@ -75,7 +75,7 @@ struct _BTL_ADAPTER {
  */
 //-------------------------------------------------------------------------
 static BTL_ADAPTER* AdappterTable[ BTL_CLIENT_MAX ];
-static u8 CommMode;
+static u32 CommMode;
 
 /*--------------------------------------------------------------------------*/
 /* Prototypes                                                               */
